requirements.txt
Pillow==10.1.0
ipython==8.18.1
google-generativeai==0.3.1
pathway 
